pyproject.toml
[tool.poetry]
name = "spotify-playlist-export"
version = "0.1.0"
description = ""
authors = ["lyingtakemura <106205052+lyingtakemura@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.1"
pytest = "^7.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
